% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_production_index_month.R
\name{get_production_index_month}
\alias{get_production_index_month}
\title{ดึงข้อมูลดัชนีผลผลิตสินค้าเกษตรรายเดือน (Monthly Production Index)}
\usage{
get_production_index_month(
  category_code = NULL,
  product_code = NULL,
  group_code = NULL,
  date = NULL,
  start_date = NULL,
  end_date = as.character(Sys.Date()),
  page = NULL,
  api_key = NULL
)
}
\arguments{
\item{category_code}{รหัสหมวดหมู่สินค้า (อ้างอิง show_index_categories() เช่น "LIVESTOCK")}

\item{product_code}{รหัสสินค้า (อ้างอิง show_index_products() เช่น "BANANA_HOM_THONG")}

\item{group_code}{รหัสกลุ่มสินค้า (อ้างอิง show_index_groups() เช่น "OIL_CROP")}

\item{date}{ค้นหาแบบระบุเดือน (รูปแบบ YYYY-MM-DD เช่น "2026-01-01")}

\item{start_date}{วันที่เริ่มต้นในการกวาดข้อมูล (รูปแบบ YYYY-MM-DD)}

\item{end_date}{วันที่สิ้นสุด (ค่าเริ่มต้นคือวันนี้)}

\item{page}{ระบุหน้าข้อมูล (ถ้าไม่ระบุ จะกวาดข้อมูลทุกหน้า)}

\item{api_key}{API Key (ถ้ามี)}
}
\value{
data.frame ข้อมูลดัชนีผลผลิต
}
\description{
ดึงข้อมูลดัชนีผลผลิตสินค้าเกษตรรายเดือน (Monthly Production Index)
}
