% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_production_index_quarter.R
\name{get_production_index_quarter}
\alias{get_production_index_quarter}
\title{ดึงข้อมูลดัชนีผลผลิตสินค้าเกษตรรายไตรมาส (Quarterly Production Index)}
\usage{
get_production_index_quarter(
  category_code = NULL,
  product_code = NULL,
  group_code = NULL,
  date = NULL,
  start_date = NULL,
  end_date = as.character(Sys.Date()),
  page = NULL,
  api_key = NULL
)
}
\arguments{
\item{category_code}{รหัสหมวดหมู่สินค้า (อ้างอิง show_index_categories())}

\item{product_code}{รหัสสินค้า (อ้างอิง show_quarter_products())}

\item{group_code}{รหัสกลุ่มสินค้า (อ้างอิง show_index_groups())}

\item{date}{ค้นหาแบบระบุวันที่ (ระบบจะแปลงเป็นไตรมาสให้อัตโนมัติ เช่น "2026-10-01" = ไตรมาส 4)}

\item{start_date}{วันที่เริ่มต้นในการกวาดข้อมูล (รูปแบบ YYYY-MM-DD)}

\item{end_date}{วันที่สิ้นสุด (ค่าเริ่มต้นคือวันนี้)}

\item{page}{ระบุหน้าข้อมูล (ถ้าไม่ระบุ จะกวาดข้อมูลทุกหน้า)}

\item{api_key}{API Key (ถ้ามี)}
}
\value{
data.frame ข้อมูลดัชนีผลผลิตรายไตรมาส
}
\description{
ดึงข้อมูลดัชนีผลผลิตสินค้าเกษตรรายไตรมาส (Quarterly Production Index)
}
