% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekly_prices.R
\name{get_weekly_prices}
\alias{get_weekly_prices}
\title{ดึงข้อมูลราคาสินค้าเกษตรรายสัปดาห์ (Weekly Prices)}
\usage{
get_weekly_prices(
  category_code = NULL,
  product_code = NULL,
  year_th = NULL,
  month = NULL,
  page = NULL,
  api_key = NULL
)
}
\arguments{
\item{category_code}{รหัสหมวดหมู่สินค้า (อ้างอิงจาก show_weekly_categories() เช่น "BUFFALO")}

\item{product_code}{รหัสสินค้า (อ้างอิงจาก show_weekly_products() เช่น "PORK_LIVE_100")}

\item{year_th}{ปี พ.ศ. (เช่น "2569") ต้องใช้คู่กับ month}

\item{month}{เดือน (เช่น "02") ต้องใช้คู่กับ year_th}

\item{page}{ระบุหน้าข้อมูล (ถ้าไม่ระบุ หรือให้เป็น NULL ระบบจะกวาดข้อมูลทุกหน้าอัตโนมัติ)}

\item{api_key}{API Key (ถ้ามี)}
}
\value{
data.frame ข้อมูลราคาสินค้าเกษตรรายสัปดาห์
}
\description{
ดึงข้อมูลราคาสินค้าเกษตรรายสัปดาห์ (Weekly Prices)
}
