% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{show_daily_categories}
\alias{show_daily_categories}
\title{แสดงรหัสหมวดสินค้ารายวัน}
\usage{
show_daily_categories()
}
\value{
data.frame แสดงรหัสและชื่อหมวดหมู่รายวัน
}
\description{
แสดงรหัสหมวดสินค้ารายวัน
}
