% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{show_index_categories}
\alias{show_index_categories}
\title{แสดงรหัสหมวดสินค้าสำหรับดัชนีผลผลิต/ราคา}
\usage{
show_index_categories()
}
\value{
data.frame แสดงรหัสและชื่อหมวดหมู่
}
\description{
แสดงรหัสหมวดสินค้าสำหรับดัชนีผลผลิต/ราคา
}
