% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{show_index_groups}
\alias{show_index_groups}
\title{แสดงรหัสกลุ่มสินค้าสำหรับดัชนีผลผลิต/ราคา}
\usage{
show_index_groups()
}
\value{
data.frame แสดงรหัสและชื่อกลุ่มสินค้า
}
\description{
แสดงรหัสกลุ่มสินค้าสำหรับดัชนีผลผลิต/ราคา
}
