% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{show_index_products}
\alias{show_index_products}
\title{แสดงรหัสสินค้าสำหรับดัชนีผลผลิต/ราคารายเดือน}
\usage{
show_index_products()
}
\value{
data.frame แสดงรหัสและชื่อสินค้า
}
\description{
แสดงรหัสสินค้าสำหรับดัชนีผลผลิต/ราคารายเดือน
}
