% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{show_quarter_products}
\alias{show_quarter_products}
\title{แสดงรหัสสินค้าสำหรับดัชนีผลผลิต/ราคารายไตรมาส}
\usage{
show_quarter_products()
}
\value{
data.frame แสดงรหัสและชื่อสินค้า
}
\description{
แสดงรหัสสินค้าสำหรับดัชนีผลผลิต/ราคารายไตรมาส
}
