% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{show_weekly_categories}
\alias{show_weekly_categories}
\title{แสดงรหัสหมวดสินค้ารายสัปดาห์/รายเดือน}
\usage{
show_weekly_categories()
}
\value{
data.frame แสดงรหัสและชื่อหมวดหมู่
}
\description{
แสดงรหัสหมวดสินค้ารายสัปดาห์/รายเดือน
}
