% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{show_weekly_products}
\alias{show_weekly_products}
\title{แสดงรหัสสินค้ารายสัปดาห์/รายเดือน}
\usage{
show_weekly_products()
}
\value{
data.frame แสดงรหัสและชื่อสินค้า
}
\description{
แสดงรหัสสินค้ารายสัปดาห์/รายเดือน
}
