% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talatThaiR-package.R
\docType{package}
\name{talatThaiR-package}
\alias{talatThaiR}
\alias{talatThaiR-package}
\title{talatThaiR: Access Thai Agricultural Statistics from NABC API}
\description{
Provides a consistent interface to access Thai agricultural statistics from the National Agricultural Big Data Center (NABC) API. Supports daily, weekly, and monthly commodity prices as well as production and price indices at monthly, quarterly, and yearly frequencies. Features smart endpoint routing, built-in product code mappings, and automatic pagination.
}
\author{
\strong{Maintainer}: Kanthanawit Jaisong \email{kanthanawit.jaisong@gmail.com}

}
\keyword{internal}
