% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_prices.R
\name{get_daily_prices}
\alias{get_daily_prices}
\title{Get daily agricultural commodity prices}
\usage{
get_daily_prices(
  category_code = NULL,
  product_code = NULL,
  date = NULL,
  start_date = NULL,
  end_date = as.character(Sys.Date()),
  api_key = NULL
)
}
\arguments{
\item{category_code}{Category code (see \code{show_daily_categories()}, e.g. "SHRIMP")}

\item{product_code}{Product code (see \code{show_daily_products()}, e.g. "LIME_XL")}

\item{date}{Fetch prices for a specific date (format: "YYYY-MM-DD")}

\item{start_date}{Filter results from this date onward (format: "YYYY-MM-DD"). Used with category_code or product_code.}

\item{end_date}{Filter results up to this date (default: today)}

\item{api_key}{API key (if required)}
}
\value{
A data.frame of daily agricultural commodity prices
}
\description{
Get daily agricultural commodity prices
}
\examples{
# select only product_category
get_daily_prices(category_code = "RICE_MALI")
# select only product_name
get_daily_prices(product_code = "LIME_XL")
# if want to know the price of all product from specific date to now
get_daily_prices(date = "2025-06-01")
# can want to know the daily price of a product from specific date to now
get_daily_prices(product_code = "LIME_XL", start_date = "2026-01-01")
can want to know the daily price of a product category from specific to now
get_daily_prices(category_code = "SHRIMP", start_date = "2026-01-01")
}
