% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_prices.R
\name{get_monthly_prices}
\alias{get_monthly_prices}
\title{ดึงข้อมูลราคาสินค้าเกษตรรายเดือน (Monthly Prices)}
\usage{
get_monthly_prices(
  category_code = NULL,
  product_code = NULL,
  year_th = NULL,
  month = NULL,
  page = NULL,
  api_key = NULL
)
}
\arguments{
\item{category_code}{รหัสหมวดหมู่สินค้า (ใช้ตารางร่วมกับรายสัปดาห์ อ้างอิง show_weekly_categories())}

\item{product_code}{รหัสสินค้า (ใช้ตารางร่วมกับรายสัปดาห์ อ้างอิง show_weekly_products())}

\item{year_th}{ปี พ.ศ. (เช่น "2569") ต้องใช้คู่กับ month}

\item{month}{เดือน (เช่น "02") ต้องใช้คู่กับ year_th}

\item{page}{ระบุหน้าข้อมูล (ถ้าไม่ระบุ หรือให้เป็น NULL ระบบจะกวาดข้อมูลทุกหน้าอัตโนมัติ)}

\item{api_key}{API Key (ถ้ามี)}
}
\value{
data.frame ข้อมูลราคาสินค้าเกษตรรายเดือน
}
\description{
ดึงข้อมูลราคาสินค้าเกษตรรายเดือน (Monthly Prices)
}
